Interim save before splitting off cmd param handling to separate project
